P O N G Beta
Finally BETA!!! I've been working on a splash menu, didn't quite make it work, but im still working on it, probably will be available in the next update.
This update mainly focuses on Making the Menu more... nice. And the Menu is WAY better than before, check it out!

USE JAVA 16
I'll try and figure out a way to be able to use this with java 8. Sorry for the inconvenience.

ALL CODE WAS MADE BY ME. :)

Enjoy!!